Return react tests and add basic ones
